Extract logic from CLI file
